start to merge schema_v4 and refactoring on Storage. all test passed up until v3.
