+++ s2core/src/main/scala/com/kakao/s2graph/core/QueryResult.scala
@@ -31,6 +31,7 @@ case class QueryRequest(query: Query,
 
 
 case class QueryResult(edgeWithScoreLs: Seq[EdgeWithScore] = Nil,
+                       tailCursor: Array[Byte] = Array.empty,
                        timestamp: Long = System.currentTimeMillis(),
                        isFailure: Boolean = false)
