+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/SKeyValue.scala
@@ -1,11 +1,28 @@
 package com.kakao.s2graph.core.storage
 
+import org.apache.hadoop.hbase.util.Bytes
 import org.hbase.async.KeyValue
 
-case class SKeyValue(table: Array[Byte], row: Array[Byte], cf: Array[Byte], qualifier: Array[Byte], value: Array[Byte], timestamp: Long) {
+
+object SKeyValue {
+  val Put = 1
+  val Delete = 2
+  val Increment = 3
+  val Default = Put
+}
+case class SKeyValue(table: Array[Byte],
+                     row: Array[Byte],
+                     cf: Array[Byte],
+                     qualifier: Array[Byte],
+                     value: Array[Byte],
+                     timestamp: Long,
+                     operation: Int = SKeyValue.Default) {
   def toLogString = {
-    Map("table" -> table.toList, "row" -> row.toList, "cf" -> cf.toList, "qualifier" -> qualifier.toList, "value" -> value.toList, "timestamp" -> timestamp).toString
+    Map("table" -> table.toList, "row" -> row.toList, "cf" -> Bytes.toString(cf),
+      "qualifier" -> qualifier.toList, "value" -> value.toList, "timestamp" -> timestamp,
+      "operation" -> operation).toString
   }
+  override def toString(): String = toLogString
 }
 
 trait CanSKeyValue[T] {
@@ -28,4 +45,3 @@ object CanSKeyValue {
 
   // For hbase KeyValues
 }
-
