+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/hbase/AsynchbaseStorage.scala
@@ -1,28 +1,26 @@
 package com.kakao.s2graph.core.storage.hbase
 
 import java.util
+import java.util.Base64
+import java.util.concurrent.TimeUnit
 
-import com.google.common.cache.Cache
-import com.kakao.s2graph.core.ExceptionHandler.{KafkaMessage, Key, Val}
-import com.kakao.s2graph.core.GraphExceptions.FetchTimeoutException
+import com.google.common.cache.{CacheBuilder}
 import com.kakao.s2graph.core._
-import com.kakao.s2graph.core.mysqls.{Label, LabelMeta}
-import com.kakao.s2graph.core.storage.Storage
+import com.kakao.s2graph.core.mysqls._
+import com.kakao.s2graph.core.storage._
 import com.kakao.s2graph.core.types._
 import com.kakao.s2graph.core.utils.{Extensions, logger}
 import com.stumbleupon.async.Deferred
-import com.typesafe.config.Config
+import com.typesafe.config.{ConfigFactory, Config}
 import org.apache.hadoop.hbase.client.{ConnectionFactory, Durability}
 import org.apache.hadoop.hbase.io.compress.Compression.Algorithm
 import org.apache.hadoop.hbase.io.encoding.DataBlockEncoding
 import org.apache.hadoop.hbase.regionserver.BloomType
 import org.apache.hadoop.hbase.util.Bytes
 import org.apache.hadoop.hbase.{HBaseConfiguration, HColumnDescriptor, HTableDescriptor, TableName}
-import org.apache.kafka.clients.producer.ProducerRecord
 import org.hbase.async._
-
 import scala.collection.JavaConversions._
-import scala.collection.Seq
+import scala.collection.{Map, Seq}
 import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, ExecutionContext, Future, duration}
 import scala.util.Random
@@ -52,59 +50,285 @@ object AsynchbaseStorage {
   }
 }
 
-class AsynchbaseStorage(override val config: Config, vertexCache: Cache[Integer, Option[Vertex]])
-                       (implicit ec: ExecutionContext) extends Storage(config) {
-
-  import AsynchbaseStorage._
-
-  //  import Extensions.FutureOps
+class AsynchbaseStorage(override val config: Config)(implicit ec: ExecutionContext)
+  extends Storage[HBaseRpc, Deferred[QueryRequestWithResult]](config) {
 
   import Extensions.DeferOps
 
+  val configWithFlush = config.withFallback(ConfigFactory.parseMap(Map("hbase.rpcs.buffered_flush_interval" -> "0")))
   val client = AsynchbaseStorage.makeClient(config)
-  val queryBuilder = new AsynchbaseQueryBuilder(this)(ec)
-  val mutationBuilder = new AsynchbaseMutationBuilder(this)(ec)
-
-  //  val cacheOpt = Option(cache)
-  val cacheOpt = None
-  val vertexCacheOpt = Option(vertexCache)
 
   private val clientWithFlush = AsynchbaseStorage.makeClient(config, "hbase.rpcs.buffered_flush_interval" -> "0")
   private val clients = Seq(client, clientWithFlush)
-
   private val clientFlushInterval = config.getInt("hbase.rpcs.buffered_flush_interval").toString().toShort
-  val MaxRetryNum = config.getInt("max.retry.number")
-  val MaxBackOff = config.getInt("max.back.off")
-  val DeleteAllFetchSize = config.getInt("delete.all.fetch.size")
-  val FailProb = config.getDouble("hbase.fail.prob")
-  val LockExpireDuration = config.getInt("lock.expire.time")
+  private val emptyKeyValues = new util.ArrayList[KeyValue]()
 
-  /**
-    * Serializer/Deserializer
-    */
-  def snapshotEdgeSerializer(snapshotEdge: SnapshotEdge) = new SnapshotEdgeSerializable(snapshotEdge)
+  private def client(withWait: Boolean): HBaseClient = if (withWait) clientWithFlush else client
 
-  def indexEdgeSerializer(indexedEdge: IndexEdge) = new IndexEdgeSerializable(indexedEdge)
+  /** Future Cache to squash request */
+  private val futureCache = CacheBuilder.newBuilder()
+  .initialCapacity(maxSize)
+  .concurrencyLevel(Runtime.getRuntime.availableProcessors())
+  .expireAfterWrite(expireAfterWrite, TimeUnit.MILLISECONDS)
+  .expireAfterAccess(expireAfterAccess, TimeUnit.MILLISECONDS)
+  .maximumSize(maxSize).build[java.lang.Long, (Long, Deferred[QueryRequestWithResult])]()
 
-  def vertexSerializer(vertex: Vertex) = new VertexSerializable(vertex)
+  /** Simple Vertex Cache */
+  private val vertexCache = CacheBuilder.newBuilder()
+  .initialCapacity(maxSize)
+  .concurrencyLevel(Runtime.getRuntime.availableProcessors())
+  .expireAfterWrite(expireAfterWrite, TimeUnit.MILLISECONDS)
+  .expireAfterAccess(expireAfterAccess, TimeUnit.MILLISECONDS)
+  .maximumSize(maxSize).build[java.lang.Integer, Option[Vertex]]()
 
-  val snapshotEdgeDeserializer = new SnapshotEdgeDeserializable
-  val indexEdgeDeserializer = new IndexEdgeDeserializable
-  val vertexDeserializer = new VertexDeserializable
 
-  def getEdges(q: Query): Future[Seq[QueryRequestWithResult]] = queryBuilder.getEdges(q)
+  override def writeToStorage(rpc: HBaseRpc, withWait: Boolean): Future[Boolean] = {
+    //    logger.debug(s"$rpc")
+    val _client = client(withWait)
+    val _defer = rpc match {
+      case d: DeleteRequest => _client.delete(d)
+      case p: PutRequest => _client.put(p)
+      case i: AtomicIncrementRequest => _client.atomicIncrement(i)
+    }
+    val future = _defer.withCallback { ret => true }.recoverWith { ex =>
+      logger.error(s"mutation failed. $rpc", ex)
+      false
+    }.toFuture
 
-  def checkEdges(params: Seq[(Vertex, Vertex, QueryParam)]): Future[Seq[QueryRequestWithResult]] = {
-    val futures = for {
-      (srcVertex, tgtVertex, queryParam) <- params
-    } yield queryBuilder.getEdge(srcVertex, tgtVertex, queryParam, false).toFuture
-
-    Future.sequence(futures)
+    if (withWait) future else Future.successful(true)
   }
 
-  def getVertices(vertices: Seq[Vertex]): Future[Seq[Vertex]] = {
+
+  /** Mutation Builder */
+  override def put(kvs: Seq[SKeyValue]): Seq[HBaseRpc] =
+    kvs.map { kv => new PutRequest(kv.table, kv.row, kv.cf, kv.qualifier, kv.value, kv.timestamp) }
+
+  override def increment(kvs: Seq[SKeyValue]): Seq[HBaseRpc] =
+    kvs.map { kv => new AtomicIncrementRequest(kv.table, kv.row, kv.cf, kv.qualifier, Bytes.toLong(kv.value)) }
+
+  override def delete(kvs: Seq[SKeyValue]): Seq[HBaseRpc] =
+    kvs.map { kv =>
+      if (kv.qualifier == null) new DeleteRequest(kv.table, kv.row, kv.cf, kv.timestamp)
+      else new DeleteRequest(kv.table, kv.row, kv.cf, kv.qualifier, kv.timestamp)
+    }
+
+
+
+  override def fetchKeyValues(hbaseRpc: AnyRef): Future[Seq[SKeyValue]] = {
+    val defer = fetchKeyValuesInner(hbaseRpc)
+    defer.toFuture.map { kvsArr =>
+      kvsArr.map { kv =>
+        implicitly[CanSKeyValue[KeyValue]].toSKeyValue(kv)
+      } toSeq
+    }
+  }
+
+
+  override def buildRequest(queryRequest: QueryRequest): AnyRef = {
+    import HSerializable._
+    val queryParam = queryRequest.queryParam
+    val label = queryParam.label
+    val edge = toRequestEdge(queryRequest)
+
+    val kv = if (queryParam.tgtVertexInnerIdOpt.isDefined) {
+      val snapshotEdge = edge.toSnapshotEdge
+      snapshotEdgeSerializer(snapshotEdge).toKeyValues.head
+      //      new GetRequest(label.hbaseTableName.getBytes, kv.row, edgeCf, kv.qualifier)
+    } else {
+      val indexedEdgeOpt = edge.edgesWithIndex.find(e => e.labelIndexSeq == queryParam.labelOrderSeq)
+      assert(indexedEdgeOpt.isDefined)
+
+      val indexedEdge = indexedEdgeOpt.get
+      indexEdgeSerializer(indexedEdge).toKeyValues.head
+    }
+
+    val (minTs, maxTs) = queryParam.duration.getOrElse((0L, Long.MaxValue))
+
+    label.schemaVersion match {
+      case HBaseType.VERSION4 if queryParam.tgtVertexInnerIdOpt.isEmpty =>
+        val scanner = client.newScanner(label.hbaseTableName.getBytes)
+        scanner.setFamily(edgeCf)
+
+        /**
+         * TODO: remove this part.
+         */
+        val indexEdgeOpt = edge.edgesWithIndex.filter(edgeWithIndex => edgeWithIndex.labelIndex.seq == queryParam.labelOrderSeq).headOption
+        val indexEdge = indexEdgeOpt.getOrElse(throw new RuntimeException(s"Can`t find index for query $queryParam"))
+
+        val srcIdBytes = VertexId.toSourceVertexId(indexEdge.srcVertex.id).bytes
+        val labelWithDirBytes = indexEdge.labelWithDir.bytes
+        val labelIndexSeqWithIsInvertedBytes = StorageSerializable.labelOrderSeqWithIsInverted(indexEdge.labelIndexSeq, isInverted = false)
+        //        val labelIndexSeqWithIsInvertedStopBytes =  StorageSerializable.labelOrderSeqWithIsInverted(indexEdge.labelIndexSeq, isInverted = true)
+        val baseKey = Bytes.add(srcIdBytes, labelWithDirBytes, Bytes.add(labelIndexSeqWithIsInvertedBytes, Array.fill(1)(edge.op)))
+        val (startKey, stopKey) =
+          if (queryParam.columnRangeFilter != null) {
+            // interval is set.
+            val _startKey = queryParam.cursorOpt match {
+              case Some(cursor) => Bytes.add(Base64.getDecoder.decode(cursor), Array.fill(1)(0))
+              case None => Bytes.add(baseKey, queryParam.columnRangeFilterMinBytes)
+            }
+            (_startKey, Bytes.add(baseKey, queryParam.columnRangeFilterMaxBytes))
+          } else {
+            /**
+             * note: since propsToBytes encode size of property map at first byte, we are sure about max value here
+             */
+            val _startKey = queryParam.cursorOpt match {
+              case Some(cursor) => Bytes.add(Base64.getDecoder.decode(cursor), Array.fill(1)(0))
+              case None => baseKey
+            }
+            (_startKey, Bytes.add(baseKey, Array.fill(1)(-1)))
+          }
+        //        logger.debug(s"[StartKey]: ${startKey.toList}")
+        //        logger.debug(s"[StopKey]: ${stopKey.toList}")
+
+        scanner.setStartKey(startKey)
+        scanner.setStopKey(stopKey)
+
+        if (queryParam.limit == Int.MinValue) logger.debug(s"MinValue: $queryParam")
+
+        scanner.setMaxVersions(1)
+        scanner.setMaxNumRows(queryParam.limit)
+        scanner.setMaxTimestamp(maxTs)
+        scanner.setMinTimestamp(minTs)
+        scanner.setRpcTimeout(queryParam.rpcTimeoutInMillis)
+        // SET option for this rpc properly.
+        scanner
+      case _ =>
+        val get =
+          if (queryParam.tgtVertexInnerIdOpt.isDefined) new GetRequest(label.hbaseTableName.getBytes, kv.row, edgeCf, kv.qualifier)
+          else new GetRequest(label.hbaseTableName.getBytes, kv.row, edgeCf)
+
+        get.maxVersions(1)
+        get.setFailfast(true)
+        get.setMaxResultsPerColumnFamily(queryParam.limit)
+        get.setRowOffsetPerColumnFamily(queryParam.offset)
+        get.setMinTimestamp(minTs)
+        get.setMaxTimestamp(maxTs)
+        get.setTimeout(queryParam.rpcTimeoutInMillis)
+
+        if (queryParam.columnRangeFilter != null) get.setFilter(queryParam.columnRangeFilter)
+
+        get
+    }
+  }
+
+  override def fetch(queryRequest: QueryRequest,
+                     prevStepScore: Double,
+                     isInnerCall: Boolean,
+                     parentEdges: Seq[EdgeWithScore]): Deferred[QueryRequestWithResult] = {
+    def fetchInner(hbaseRpc: AnyRef) = {
+      fetchKeyValuesInner(hbaseRpc).withCallback { kvs =>
+
+        val edgeWithScores = toEdges(kvs, queryRequest.queryParam, prevStepScore, isInnerCall, parentEdges)
+        val resultEdgesWithScores = if (queryRequest.queryParam.sample >= 0) {
+          sample(queryRequest, edgeWithScores, queryRequest.queryParam.sample)
+        } else edgeWithScores
+        QueryRequestWithResult(queryRequest, QueryResult(resultEdgesWithScores, tailCursor = kvs.lastOption.map(_.key).getOrElse(Array.empty)))
+
+      } recoverWith { ex =>
+        logger.error(s"fetchQueryParam failed. fallback return.", ex)
+        QueryRequestWithResult(queryRequest, QueryResult(isFailure = true))
+      }
+    }
+    def checkAndExpire(hbaseRpc: AnyRef,
+                       cacheKey: Long,
+                       cacheTTL: Long,
+                       cachedAt: Long,
+                       defer: Deferred[QueryRequestWithResult]): Deferred[QueryRequestWithResult] = {
+      if (System.currentTimeMillis() >= cachedAt + cacheTTL) {
+        // future is too old. so need to expire and fetch new data from storage.
+        futureCache.asMap().remove(cacheKey)
+        val newPromise = new Deferred[QueryRequestWithResult]()
+        futureCache.asMap().putIfAbsent(cacheKey, (System.currentTimeMillis(), newPromise)) match {
+          case null =>
+            // only one thread succeed to come here concurrently
+            // initiate fetch to storage then add callback on complete to finish promise.
+            fetchInner(hbaseRpc) withCallback { queryRequestWithResult =>
+              newPromise.callback(queryRequestWithResult)
+              queryRequestWithResult
+            }
+            newPromise
+          case (cachedAt, oldDefer) => oldDefer
+        }
+      } else {
+        // future is not to old so reuse it.
+        defer
+      }
+    }
+
+    val queryParam = queryRequest.queryParam
+    val cacheTTL = queryParam.cacheTTLInMillis
+    val request = buildRequest(queryRequest)
+
+    if (cacheTTL <= 0) fetchInner(request)
+    else {
+      val cacheKeyBytes = Bytes.add(queryRequest.query.cacheKeyBytes, toCacheKeyBytes(request))
+      val cacheKey = queryParam.toCacheKey(cacheKeyBytes)
+
+      val cacheVal = futureCache.getIfPresent(cacheKey)
+      cacheVal match {
+        case null =>
+          // here there is no promise set up for this cacheKey so we need to set promise on future cache.
+          val promise = new Deferred[QueryRequestWithResult]()
+          val now = System.currentTimeMillis()
+          val (cachedAt, defer) = futureCache.asMap().putIfAbsent(cacheKey, (now, promise)) match {
+            case null =>
+              fetchInner(request) withCallback { queryRequestWithResult =>
+                promise.callback(queryRequestWithResult)
+                queryRequestWithResult
+              }
+              (now, promise)
+            case oldVal => oldVal
+          }
+          checkAndExpire(request, cacheKey, cacheTTL, cachedAt, defer)
+        case (cachedAt, defer) =>
+          checkAndExpire(request, cacheKey, cacheTTL, cachedAt, defer)
+      }
+    }
+  }
+
+
+  override def fetches(queryRequestWithScoreLs: scala.Seq[(QueryRequest, Double)],
+                       prevStepEdges: Predef.Map[VertexId, scala.Seq[EdgeWithScore]]): Future[scala.Seq[QueryRequestWithResult]] = {
+    val defers: Seq[Deferred[QueryRequestWithResult]] = for {
+      (queryRequest, prevStepScore) <- queryRequestWithScoreLs
+      parentEdges <- prevStepEdges.get(queryRequest.vertex.id)
+    } yield fetch(queryRequest, prevStepScore, isInnerCall = false, parentEdges)
+
+    val grouped: Deferred[util.ArrayList[QueryRequestWithResult]] = Deferred.group(defers)
+    grouped withCallback {
+      queryResults: util.ArrayList[QueryRequestWithResult] =>
+        queryResults.toIndexedSeq
+    } toFuture
+  }
+
+  override def incrementCounts(edges: Seq[Edge], withWait: Boolean): Future[Seq[(Boolean, Long)]] = {
+    val defers: Seq[Deferred[(Boolean, Long)]] = for {
+      edge <- edges
+    } yield {
+        val edgeWithIndex = edge.edgesWithIndex.head
+        val countWithTs = edge.propsWithTs(LabelMeta.countSeq)
+        val countVal = countWithTs.innerVal.toString().toLong
+        val incr = buildIncrementsCountAsync(edgeWithIndex, countVal).head
+        val request = incr.asInstanceOf[AtomicIncrementRequest]
+        client.bufferAtomicIncrement(request) withCallback { resultCount: java.lang.Long =>
+          (true, resultCount.longValue())
+        } recoverWith { ex =>
+          logger.error(s"mutation failed. $request", ex)
+          (false, -1L)
+        }
+      }
+
+    val grouped: Deferred[util.ArrayList[(Boolean, Long)]] = Deferred.groupInOrder(defers)
+    grouped.toFuture.map(_.toSeq)
+  }
+
+
+
+  /** Asynchbase implementation override default getVertices to use future Cache */
+  override def getVertices(vertices: Seq[Vertex]): Future[Seq[Vertex]] = {
     def fromResult(queryParam: QueryParam,
-                   kvs: Seq[org.hbase.async.KeyValue],
+                   kvs: Seq[SKeyValue],
                    version: String): Option[Vertex] = {
 
       if (kvs.isEmpty) None
@@ -116,7 +340,7 @@ class AsynchbaseStorage(override val config: Config, vertexCache: Cache[Integer,
 
     val futures = vertices.map { vertex =>
       val kvs = vertexSerializer(vertex).toKeyValues
-      val get = new GetRequest(vertex.hbaseTableName.getBytes, kvs.head.row, vertexCf)
+      val get = new GetRequest(vertex.hbaseTableName.getBytes, kvs.head.row, HSerializable.vertexCf)
       //      get.setTimeout(this.singleGetTimeout.toShort)
       get.setFailfast(true)
       get.maxVersions(1)
@@ -124,8 +348,10 @@ class AsynchbaseStorage(override val config: Config, vertexCache: Cache[Integer,
       val cacheKey = MurmurHash3.stringHash(get.toString)
       val cacheVal = vertexCache.getIfPresent(cacheKey)
       if (cacheVal == null)
-        client.get(get).toFutureWith(emptyKVs).map { kvs =>
+        fetchKeyValues(get).map { kvs =>
           fromResult(QueryParam.Empty, kvs, vertex.serviceColumn.schemaVersion)
+        } recoverWith { case ex: Throwable =>
+          Future.successful(None)
         }
 
       else Future.successful(cacheVal)
@@ -134,675 +360,18 @@ class AsynchbaseStorage(override val config: Config, vertexCache: Cache[Integer,
     Future.sequence(futures).map { result => result.toList.flatten }
   }
 
-
-  def mutateEdge(edge: Edge, withWait: Boolean): Future[Boolean] = {
-    val edgeFuture =
-      if (edge.op == GraphUtil.operations("deleteAll")) {
-        deleteAllAdjacentEdges(Seq(edge.srcVertex), Seq(edge.label), edge.labelWithDir.dir, edge.ts)
-      } else {
-        val strongConsistency = edge.label.consistencyLevel == "strong"
-        if (edge.op == GraphUtil.operations("delete") && !strongConsistency) {
-          val zkQuorum = edge.label.hbaseZkAddr
-          val (_, edgeUpdate) = Edge.buildDeleteBulk(None, edge)
-          val mutations =
-            mutationBuilder.indexedEdgeMutations(edgeUpdate) ++
-              mutationBuilder.snapshotEdgeMutations(edgeUpdate) ++
-              mutationBuilder.increments(edgeUpdate)
-          writeAsyncSimple(zkQuorum, mutations, withWait)
-        } else {
-          mutateEdgesInner(Seq(edge), strongConsistency, withWait)(Edge.buildOperation)
-        }
-      }
-
-    val vertexFuture = writeAsyncSimple(edge.label.hbaseZkAddr,
-      mutationBuilder.buildVertexPutsAsync(edge), withWait)
-
-    Future.sequence(Seq(edgeFuture, vertexFuture)).map(_.forall(identity))
-  }
-
-  override def mutateEdges(_edges: Seq[Edge], withWait: Boolean): Future[Seq[Boolean]] = {
-    val grouped = _edges.groupBy { edge => (edge.label, edge.srcVertex.innerId, edge.tgtVertex.innerId) } toSeq
-
-    val mutateEdges = grouped.map { case ((_, _, _), edgeGroup) =>
-      val (deleteAllEdges, edges) = edgeGroup.partition(_.op == GraphUtil.operations("deleteAll"))
-
-      // DeleteAll first
-      val deleteAllFutures = deleteAllEdges.map { edge =>
-        deleteAllAdjacentEdges(Seq(edge.srcVertex), Seq(edge.label), edge.labelWithDir.dir, edge.ts)
-      }
-
-      // After deleteAll, process others
-      lazy val mutateEdgeFutures = edges.toList match {
-        case head :: tail =>
-          val strongConsistency = edges.head.label.consistencyLevel == "strong"
-          if (strongConsistency) {
-            val edgeFuture = mutateEdgesInner(edges, strongConsistency, withWait)(Edge.buildOperation)
-
-            //TODO: decide what we will do on failure on vertex put
-            val puts = mutationBuilder.buildVertexPutsAsync(head)
-            val vertexFuture = writeAsyncSimple(head.label.hbaseZkAddr, puts, withWait)
-            Seq(edgeFuture, vertexFuture)
-          } else {
-            edges.map { edge => mutateEdge(edge, withWait = withWait) }
-          }
-        case Nil => Nil
-      }
-
-      val composed = for {
-        deleteRet <- Future.sequence(deleteAllFutures)
-        mutateRet <- Future.sequence(mutateEdgeFutures)
-      } yield deleteRet ++ mutateRet
-
-      composed.map(_.forall(identity))
-    }
-
-    Future.sequence(mutateEdges)
-  }
-
-  def mutateVertex(vertex: Vertex, withWait: Boolean): Future[Boolean] = {
-    if (vertex.op == GraphUtil.operations("delete")) {
-      writeAsyncSimple(vertex.hbaseZkAddr, mutationBuilder.buildDeleteAsync(vertex), withWait)
-    } else if (vertex.op == GraphUtil.operations("deleteAll")) {
-      logger.info(s"deleteAll for vertex is truncated. $vertex")
-      Future.successful(true) // Ignore withWait parameter, because deleteAll operation may takes long time
-    } else {
-      writeAsyncSimple(vertex.hbaseZkAddr, mutationBuilder.buildPutsAll(vertex), withWait)
-    }
-  }
-
-  def incrementCounts(edges: Seq[Edge], withWait: Boolean): Future[Seq[(Boolean, Long)]] = {
-    val _client = if (withWait) clientWithFlush else client
-    val defers: Seq[Deferred[(Boolean, Long)]] = for {
-      edge <- edges
-    } yield {
-      val edgeWithIndex = edge.edgesWithIndex.head
-      val countWithTs = edge.propsWithTs(LabelMeta.countSeq)
-      val countVal = countWithTs.innerVal.toString().toLong
-      val incr = mutationBuilder.buildIncrementsCountAsync(edgeWithIndex, countVal).head
-      val request = incr.asInstanceOf[AtomicIncrementRequest]
-      val defer = _client.bufferAtomicIncrement(request) withCallback { resultCount: java.lang.Long =>
-        (true, resultCount.longValue())
-      } recoverWith { ex =>
-        logger.error(s"mutation failed. $request", ex)
-        (false, -1L)
-      }
-      if (withWait) defer
-      else Deferred.fromResult((true, -1L))
-    }
-
-    val grouped: Deferred[util.ArrayList[(Boolean, Long)]] = Deferred.groupInOrder(defers)
-    grouped.toFuture.map(_.toSeq)
-  }
-
-  private def writeAsyncSimpleRetry(zkQuorum: String, elementRpcs: Seq[HBaseRpc], withWait: Boolean): Future[Boolean] = {
-    def compute = writeAsyncSimple(zkQuorum, elementRpcs, withWait).flatMap { ret =>
-      if (ret) Future.successful(ret)
-      else throw FetchTimeoutException("writeAsyncWithWaitRetrySimple")
-    }
-    Extensions.retryOnFailure(MaxRetryNum) {
-      compute
-    } {
-      logger.error(s"writeAsyncWithWaitRetrySimple: $elementRpcs")
-      false
-    }
-  }
-
-  private def writeToStorage(_client: HBaseClient, rpc: HBaseRpc): Deferred[Boolean] = {
-    //    logger.debug(s"$rpc")
-    val defer = rpc match {
-      case d: DeleteRequest => _client.delete(d)
-      case p: PutRequest => _client.put(p)
-      case i: AtomicIncrementRequest => _client.bufferAtomicIncrement(i)
-    }
-    defer withCallback { ret => true } recoverWith { ex =>
-      logger.error(s"mutation failed. $rpc", ex)
-      false
-    }
-  }
-
-  private def writeAsyncSimple(zkQuorum: String, elementRpcs: Seq[HBaseRpc], withWait: Boolean): Future[Boolean] = {
-    val _client = if (withWait) clientWithFlush else client
-    if (elementRpcs.isEmpty) {
-      Future.successful(true)
-    } else {
-      val defers = elementRpcs.map { rpc => writeToStorage(_client, rpc) }
-      if (withWait)
-        Deferred.group(defers).toFuture map { arr => arr.forall(identity) }
-      else
-        Future.successful(true)
-    }
-  }
-
-  private def writeAsync(zkQuorum: String, elementRpcs: Seq[Seq[HBaseRpc]], withWait: Boolean): Future[Seq[Boolean]] = {
-    val _client = if (withWait) clientWithFlush else client
-    if (elementRpcs.isEmpty) {
-      Future.successful(Seq.empty[Boolean])
-    } else {
-      val futures = elementRpcs.map { rpcs =>
-        val defers = rpcs.map { rpc => writeToStorage(_client, rpc) }
-        if (withWait)
-          Deferred.group(defers).toFuture map { arr => arr.forall(identity) }
-        else
-          Future.successful(true)
-      }
-      if (withWait)
-        Future.sequence(futures)
-      else
-        Future.successful(elementRpcs.map(_ => true))
-    }
-  }
-
-  private def fetchSnapshotEdge(edge: Edge): Future[(QueryParam, Option[Edge], Option[KeyValue])] = {
-    val labelWithDir = edge.labelWithDir
-    val queryParam = QueryParam(labelWithDir)
-    val _queryParam = queryParam.tgtVertexInnerIdOpt(Option(edge.tgtVertex.innerId))
-    val q = Query.toQuery(Seq(edge.srcVertex), _queryParam)
-    val queryRequest = QueryRequest(q, 0, edge.srcVertex, _queryParam)
-
-    client.get(queryBuilder.buildRequest(queryRequest)) withCallback { kvs =>
-      val (edgeOpt, kvOpt) =
-        if (kvs.isEmpty()) (None, None)
-        else {
-          val _edgeOpt = toEdges(kvs, queryParam, 1.0, isInnerCall = true, parentEdges = Nil).headOption.map(_.edge)
-          val _kvOpt = kvs.headOption
-          (_edgeOpt, _kvOpt)
-        }
-      (queryParam, edgeOpt, kvOpt)
-    } recoverWith { ex =>
-      logger.error(s"fetchQueryParam failed. fallback return.", ex)
-      throw new FetchTimeoutException(s"${edge.toLogString}")
-    } toFuture
-  }
-
-
-  case class PartialFailureException(edge: Edge, statusCode: Byte, faileReason: String) extends Exception
-
-  def debug(ret: Boolean, phase: String, snapshotEdge: SnapshotEdge) = {
-    val msg = Seq(s"[$ret] [$phase]", s"${snapshotEdge.toLogString()}").mkString("\n")
-    logger.debug(msg)
-  }
-
-  def debug(ret: Boolean, phase: String, snapshotEdge: SnapshotEdge, edgeMutate: EdgeMutate) = {
-    val msg = Seq(s"[$ret] [$phase]", s"${snapshotEdge.toLogString()}",
-      s"${edgeMutate.toLogString}").mkString("\n")
-    logger.debug(msg)
-  }
-
-  private def buildLockEdge(snapshotEdgeOpt: Option[Edge], edge: Edge, kvOpt: Option[KeyValue]) = {
-    val currentTs = System.currentTimeMillis()
-    val lockTs = snapshotEdgeOpt match {
-      case None => Option(currentTs)
-      case Some(snapshotEdge) =>
-        snapshotEdge.pendingEdgeOpt match {
-          case None => Option(currentTs)
-          case Some(pendingEdge) => pendingEdge.lockTs
-        }
-    }
-    val newVersion = kvOpt.map(_.timestamp()).getOrElse(edge.ts) + 1
-    //      snapshotEdgeOpt.map(_.version).getOrElse(edge.ts) + 1
-    val pendingEdge = edge.copy(version = newVersion, statusCode = 1, lockTs = lockTs)
-    val base = snapshotEdgeOpt match {
-      case None =>
-        // no one ever mutated on this snapshotEdge.
-        edge.toSnapshotEdge.copy(pendingEdgeOpt = Option(pendingEdge))
-      case Some(snapshotEdge) =>
-        // there is at least one mutation have been succeed.
-        snapshotEdgeOpt.get.toSnapshotEdge.copy(pendingEdgeOpt = Option(pendingEdge))
-    }
-    base.copy(version = newVersion, statusCode = 1, lockTs = None)
-  }
-
-  private def buildReleaseLockEdge(snapshotEdgeOpt: Option[Edge], lockEdge: SnapshotEdge,
-                                   edgeMutate: EdgeMutate) = {
-    val newVersion = lockEdge.version + 1
-    val base = edgeMutate.newSnapshotEdge match {
-      case None =>
-        // shouldReplace false
-        assert(snapshotEdgeOpt.isDefined)
-        snapshotEdgeOpt.get.toSnapshotEdge
-      case Some(newSnapshotEdge) => newSnapshotEdge
-    }
-    base.copy(version = newVersion, statusCode = 0, pendingEdgeOpt = None)
-  }
-
-  def mutate(predicate: Boolean,
-             edge: Edge,
-             statusCode: Byte,
-             _edgeMutate: EdgeMutate): Future[Boolean] = {
-    if (!predicate) throw new PartialFailureException(edge, 1, "predicate failed.")
-
-    if (statusCode >= 2) {
-      logger.debug(s"skip mutate: [$statusCode]\n${edge.toLogString}")
-      Future.successful(true)
-    } else {
-      val p = Random.nextDouble()
-      if (p < FailProb) throw new PartialFailureException(edge, 1, s"$p")
-      else
-        writeAsyncSimple(edge.label.hbaseZkAddr, mutationBuilder.indexedEdgeMutations(_edgeMutate), withWait = true).map { ret =>
-          if (ret) {
-            debug(ret, "mutate", edge.toSnapshotEdge, _edgeMutate)
-          } else {
-            throw new PartialFailureException(edge, 1, "hbase fail.")
-          }
-          true
-        }
-    }
-  }
-
-  def increment(predicate: Boolean,
-                edge: Edge,
-                statusCode: Byte, _edgeMutate: EdgeMutate): Future[Boolean] = {
-    if (!predicate) throw new PartialFailureException(edge, 2, "predicate failed.")
-    if (statusCode >= 3) {
-      logger.debug(s"skip increment: [$statusCode]\n${edge.toLogString}")
-      Future.successful(true)
-    } else {
-      val p = Random.nextDouble()
-      if (p < FailProb) throw new PartialFailureException(edge, 2, s"$p")
-      else
-        writeAsyncSimple(edge.label.hbaseZkAddr, mutationBuilder.increments(_edgeMutate), withWait = true).map { ret =>
-          if (ret) {
-            debug(ret, "increment", edge.toSnapshotEdge, _edgeMutate)
-          } else {
-            throw new PartialFailureException(edge, 2, "hbase fail.")
-          }
-          true
-        }
-    }
-  }
-
-  def acquireLock(statusCode: Byte, edge: Edge,
-                  lockEdge: SnapshotEdge, oldBytes: Array[Byte]): Future[Boolean] =
-    if (statusCode >= 1) {
-      logger.debug(s"skip acquireLock: [$statusCode]\n${edge.toLogString}")
-      Future.successful(true)
-    } else {
-      val p = Random.nextDouble()
-      if (p < FailProb) throw new PartialFailureException(edge, 0, s"$p")
-      else {
-        val lockEdgePut = toPutRequest(lockEdge)
-        client.compareAndSet(lockEdgePut, oldBytes).toFuture.recoverWith {
-          case ex: Exception =>
-            logger.error(s"AcquireLock RPC Failed.")
-            throw new PartialFailureException(edge, 0, "AcquireLock RPC Failed")
-        }.map { ret =>
-          if (ret) {
-            val log = Seq(
-              "\n",
-              "=" * 50,
-              s"[Success]: acquireLock",
-              s"[RequestEdge]: ${edge.toLogString}",
-              s"[LockEdge]: ${lockEdge.toLogString()}",
-              s"[PendingEdge]: ${lockEdge.pendingEdgeOpt.map(_.toLogString).getOrElse("")}",
-              "=" * 50, "\n").mkString("\n")
-
-            logger.debug(log)
-            //            debug(ret, "acquireLock", edge.toSnapshotEdge)
-          } else {
-            throw new PartialFailureException(edge, 0, "hbase fail.")
-          }
-          true
-        }
-      }
-    }
-
-  def releaseLock(predicate: Boolean,
-                  edge: Edge,
-                  lockEdge: SnapshotEdge,
-                  releaseLockEdge: SnapshotEdge,
-                  _edgeMutate: EdgeMutate,
-                  oldBytes: Array[Byte]): Future[Boolean] = {
-    if (!predicate) {
-      throw new PartialFailureException(edge, 3, "predicate failed.")
-    }
-    val p = Random.nextDouble()
-    if (p < FailProb) throw new PartialFailureException(edge, 3, s"$p")
-    else {
-      val releaseLockEdgePut = toPutRequest(releaseLockEdge)
-      val lockEdgePut = toPutRequest(lockEdge)
-
-      client.compareAndSet(releaseLockEdgePut, lockEdgePut.value()).toFuture.recoverWith {
-        case ex: Exception =>
-          logger.error(s"ReleaseLock RPC Failed.")
-          throw new PartialFailureException(edge, 3, "ReleaseLock RPC Failed")
-      }.map { ret =>
-        if (ret) {
-          debug(ret, "releaseLock", edge.toSnapshotEdge)
-        } else {
-          val msg = Seq("\nFATAL ERROR\n",
-            "=" * 50,
-            oldBytes.toList,
-            lockEdgePut.value.toList,
-            releaseLockEdgePut.value().toList,
-            "=" * 50,
-            "\n"
-          )
-          logger.error(msg.mkString("\n"))
-          //          error(ret, "releaseLock", edge.toSnapshotEdge)
-          throw new PartialFailureException(edge, 3, "hbase fail.")
-        }
-        true
-      }
-    }
-    //      }
-  }
-
-  private def toPutRequest(snapshotEdge: SnapshotEdge): PutRequest = {
-    mutationBuilder.buildPutAsync(snapshotEdge).head.asInstanceOf[PutRequest]
-  }
-
-
-  private def commitUpdate(edge: Edge,
-                           statusCode: Byte)(snapshotEdgeOpt: Option[Edge],
-                                             kvOpt: Option[KeyValue],
-                                             edgeUpdate: EdgeMutate): Future[Boolean] = {
-    val label = edge.label
-    def oldBytes = kvOpt.map(_.value()).getOrElse(Array.empty)
-    //    def oldBytes = snapshotEdgeOpt.map { e =>
-    //      snapshotEdgeSerializer(e.toSnapshotEdge).toKeyValues.head.value
-    //    }.getOrElse(Array.empty)
-    def process(lockEdge: SnapshotEdge,
-                releaseLockEdge: SnapshotEdge,
-                _edgeMutate: EdgeMutate,
-                statusCode: Byte): Future[Boolean] = {
-
-      for {
-        locked <- acquireLock(statusCode, edge, lockEdge, oldBytes)
-        mutated <- mutate(locked, edge, statusCode, _edgeMutate)
-        incremented <- increment(mutated, edge, statusCode, _edgeMutate)
-        released <- releaseLock(incremented, edge, lockEdge, releaseLockEdge, _edgeMutate, oldBytes)
-      } yield {
-        released
-      }
-    }
-
-
-    val lockEdge = buildLockEdge(snapshotEdgeOpt, edge, kvOpt)
-    val releaseLockEdge = buildReleaseLockEdge(snapshotEdgeOpt, lockEdge, edgeUpdate)
-    snapshotEdgeOpt match {
-      case None =>
-        // no one ever did success on acquire lock.
-        process(lockEdge, releaseLockEdge, edgeUpdate, statusCode)
-      case Some(snapshotEdge) =>
-        // someone did success on acquire lock at least one.
-        snapshotEdge.pendingEdgeOpt match {
-          case None =>
-            // not locked
-            process(lockEdge, releaseLockEdge, edgeUpdate, statusCode)
-          case Some(pendingEdge) =>
-            def isLockExpired = pendingEdge.lockTs.get + LockExpireDuration < System.currentTimeMillis()
-            if (isLockExpired) {
-              val oldSnapshotEdge = if (snapshotEdge.ts == pendingEdge.ts) None else Option(snapshotEdge)
-              val (_, newEdgeUpdate) = Edge.buildOperation(oldSnapshotEdge, Seq(pendingEdge))
-              val newLockEdge = buildLockEdge(snapshotEdgeOpt, pendingEdge, kvOpt)
-              val _newReleaseLockEdge = buildReleaseLockEdge(snapshotEdgeOpt, newLockEdge, newEdgeUpdate)
-
-              // set lock ts as current ts
-              val newPendingEdgeOpt = _newReleaseLockEdge.pendingEdgeOpt.map(_.copy(lockTs = Option(System.currentTimeMillis())))
-              val newReleaseLockEdge = _newReleaseLockEdge.copy(pendingEdgeOpt = newPendingEdgeOpt)
-
-              process(newLockEdge, newReleaseLockEdge, newEdgeUpdate, statusCode = 0).flatMap { ret =>
-                val log = s"[Success]: Resolving expired pending edge.\n${pendingEdge.toLogString}"
-                throw new PartialFailureException(edge, 0, log)
-              }
-            } else {
-              // locked
-              if (pendingEdge.ts == edge.ts && statusCode > 0) {
-                // self locked
-                val oldSnapshotEdge = if (snapshotEdge.ts == pendingEdge.ts) None else Option(snapshotEdge)
-                val (_, newEdgeUpdate) = Edge.buildOperation(oldSnapshotEdge, Seq(edge))
-                val newReleaseLockEdge = buildReleaseLockEdge(snapshotEdgeOpt, lockEdge, newEdgeUpdate)
-
-                /** lockEdge will be ignored */
-                process(lockEdge, newReleaseLockEdge, newEdgeUpdate, statusCode)
-              } else {
-                throw new PartialFailureException(edge, statusCode, s"others[${pendingEdge.ts}] is mutating. me[${edge.ts}]")
-              }
-            }
-        }
-    }
-  }
-
-  private def mutateEdgesInner(edges: Seq[Edge],
-                               checkConsistency: Boolean,
-                               withWait: Boolean)(f: (Option[Edge], Seq[Edge]) => (Edge, EdgeMutate)): Future[Boolean] = {
-    if (!checkConsistency) {
-      val zkQuorum = edges.head.label.hbaseZkAddr
-      val futures = edges.map { edge =>
-        val (_, edgeUpdate) = f(None, Seq(edge))
-        val mutations =
-          mutationBuilder.indexedEdgeMutations(edgeUpdate) ++
-            mutationBuilder.snapshotEdgeMutations(edgeUpdate) ++
-            mutationBuilder.increments(edgeUpdate)
-        writeAsyncSimple(zkQuorum, mutations, withWait)
-      }
-      Future.sequence(futures).map { rets => rets.forall(identity) }
-    } else {
-      def commit(_edges: Seq[Edge], statusCode: Byte): Future[Boolean] = {
-
-        fetchSnapshotEdge(_edges.head) flatMap { case (queryParam, snapshotEdgeOpt, kvOpt) =>
-
-          val (newEdge, edgeUpdate) = f(snapshotEdgeOpt, _edges)
-          //shouldReplace false.
-          if (edgeUpdate.newSnapshotEdge.isEmpty && statusCode <= 0) {
-            logger.debug(s"${newEdge.toLogString} drop.")
-            Future.successful(true)
-          } else {
-            commitUpdate(newEdge, statusCode)(snapshotEdgeOpt, kvOpt, edgeUpdate).map { ret =>
-              if (ret) {
-                logger.info(s"[Success] commit: \n${_edges.map(_.toLogString).mkString("\n")}")
-              } else {
-                throw new PartialFailureException(newEdge, 3, "commit failed.")
-              }
-              true
-            }
-          }
-        }
-      }
-      def retry(tryNum: Int)(edges: Seq[Edge], statusCode: Byte)(fn: (Seq[Edge], Byte) => Future[Boolean]): Future[Boolean] = {
-        if (tryNum >= MaxRetryNum) {
-          edges.foreach { edge =>
-            logger.error(s"commit failed after $MaxRetryNum\n${edge.toLogString}")
-            ExceptionHandler.enqueue(ExceptionHandler.toKafkaMessage(element = edge))
-          }
-          Future.successful(false)
-        } else {
-          val future = fn(edges, statusCode)
-          future.onSuccess {
-            case success =>
-              logger.debug(s"Finished. [$tryNum]\n${edges.head.toLogString}\n")
-          }
-          future recoverWith {
-            case FetchTimeoutException(retryEdge) =>
-              logger.info(s"[Try: $tryNum], Fetch fail.\n${retryEdge}")
-              retry(tryNum + 1)(edges, statusCode)(fn)
-
-            case PartialFailureException(retryEdge, failedStatusCode, faileReason) =>
-              val status = failedStatusCode match {
-                case 0 => "AcquireLock failed."
-                case 1 => "Mutation failed."
-                case 2 => "Increment failed."
-                case 3 => "ReleaseLock failed."
-                case 4 => "Unknown"
-              }
-
-              Thread.sleep(Random.nextInt(MaxBackOff))
-              logger.info(s"[Try: $tryNum], [Status: $status] partial fail.\n${retryEdge.toLogString}\nFailReason: ${faileReason}")
-              retry(tryNum + 1)(Seq(retryEdge), failedStatusCode)(fn)
-            case ex: Exception =>
-              logger.error("Unknown exception", ex)
-              Future.successful(false)
-          }
-        }
-      }
-      retry(1)(edges, 0)(commit)
-    }
-  }
-
-
-  def mutateLog(snapshotEdgeOpt: Option[Edge], edges: Seq[Edge],
-                newEdge: Edge, edgeMutate: EdgeMutate) = {
-    Seq("----------------------------------------------",
-      s"SnapshotEdge: ${snapshotEdgeOpt.map(_.toLogString)}",
-      s"requestEdges: ${edges.map(_.toLogString).mkString("\n")}",
-      s"newEdge: ${newEdge.toLogString}",
-      s"mutation: \n${edgeMutate.toLogString}",
-      "----------------------------------------------").mkString("\n")
-  }
-
-  private def deleteAllFetchedEdgesAsyncOld(queryRequest: QueryRequest,
-                                            queryResult: QueryResult,
-                                            requestTs: Long,
-                                            retryNum: Int): Future[Boolean] = {
-    val queryParam = queryRequest.queryParam
-    val zkQuorum = queryParam.label.hbaseZkAddr
-    val futures = for {
-      edgeWithScore <- queryResult.edgeWithScoreLs
-      (edge, score) = EdgeWithScore.unapply(edgeWithScore).get
-    } yield {
-      /** reverted direction */
-      val reversedIndexedEdgesMutations = edge.duplicateEdge.edgesWithIndex.flatMap { indexedEdge =>
-        mutationBuilder.buildDeletesAsync(indexedEdge) ++ mutationBuilder.buildIncrementsAsync(indexedEdge, -1L)
-      }
-      val reversedSnapshotEdgeMutations = mutationBuilder.buildDeleteAsync(edge.toSnapshotEdge)
-      val forwardIndexedEdgeMutations = edge.edgesWithIndex.flatMap { indexedEdge =>
-        mutationBuilder.buildDeletesAsync(indexedEdge) ++ mutationBuilder.buildIncrementsAsync(indexedEdge, -1L)
-      }
-      val mutations = reversedIndexedEdgesMutations ++ reversedSnapshotEdgeMutations ++ forwardIndexedEdgeMutations
-      writeAsyncSimple(zkQuorum, mutations, withWait = true)
-    }
-
-    Future.sequence(futures).map { rets => rets.forall(identity) }
-  }
-
-  private def buildEdgesToDelete(queryRequestWithResultLs: QueryRequestWithResult, requestTs: Long): QueryResult = {
-    val (queryRequest, queryResult) = QueryRequestWithResult.unapply(queryRequestWithResultLs).get
-    val edgeWithScoreLs = queryResult.edgeWithScoreLs.filter { edgeWithScore =>
-      (edgeWithScore.edge.ts < requestTs) && !edgeWithScore.edge.isDegree
-    }.map { edgeWithScore =>
-      val label = queryRequest.queryParam.label
-      val newPropsWithTs = edgeWithScore.edge.propsWithTs ++
-        Map(LabelMeta.timeStampSeq -> InnerValLikeWithTs.withLong(requestTs, requestTs, label.schemaVersion))
-      val copiedEdge = edgeWithScore.edge.copy(op = GraphUtil.operations("delete"), version = requestTs,
-        propsWithTs = newPropsWithTs)
-      edgeWithScore.copy(edge = copiedEdge)
-    }
-    queryResult.copy(edgeWithScoreLs = edgeWithScoreLs)
-  }
-
-  private def deleteAllFetchedEdgesLs(queryRequestWithResultLs: Seq[QueryRequestWithResult], requestTs: Long): Future[(Boolean, Boolean)] = {
-    val queryResultLs = queryRequestWithResultLs.map(_.queryResult)
-    queryResultLs.foreach { queryResult =>
-      if (queryResult.isFailure) throw new RuntimeException("fetched result is fallback.")
-    }
-
-    val futures = for {
-      queryRequestWithResult <- queryRequestWithResultLs
-      (queryRequest, _) = QueryRequestWithResult.unapply(queryRequestWithResult).get
-      deleteQueryResult = buildEdgesToDelete(queryRequestWithResult, requestTs)
-      if deleteQueryResult.edgeWithScoreLs.nonEmpty
-    } yield {
-      val label = queryRequest.queryParam.label
-      label.schemaVersion match {
-        case HBaseType.VERSION3 if label.consistencyLevel == "strong" =>
-
-          /**
-            * read: snapshotEdge on queryResult = O(N)
-            * write: N x (relatedEdges x indices(indexedEdge) + 1(snapshotEdge))
-            */
-          mutateEdges(deleteQueryResult.edgeWithScoreLs.map(_.edge), withWait = true).map(_.forall(identity))
-        case _ =>
-
-          /**
-            * read: x
-            * write: N x ((1(snapshotEdge) + 2(1 for incr, 1 for delete) x indices)
-            */
-          deleteAllFetchedEdgesAsyncOld(queryRequest, deleteQueryResult, requestTs, MaxRetryNum)
-      }
-    }
-    if (futures.isEmpty) {
-      // all deleted.
-      Future.successful(true -> true)
-    } else {
-      Future.sequence(futures).map { rets => false -> rets.forall(identity) }
-    }
-  }
-
-  def fetchAndDeleteAll(query: Query, requestTs: Long): Future[(Boolean, Boolean)] = {
-    val future = for {
-      queryRequestWithResultLs <- getEdges(query)
-      (allDeleted, ret) <- deleteAllFetchedEdgesLs(queryRequestWithResultLs, requestTs)
-    } yield {
-      (allDeleted, ret)
-    }
-    Extensions.retryOnFailure(MaxRetryNum) {
-      future
-    } {
-      logger.error(s"fetch and deleteAll failed.")
-      (true, false)
-    }
-
-  }
-
-  def deleteAllAdjacentEdges(srcVertices: Seq[Vertex],
-                             labels: Seq[Label],
-                             dir: Int,
-                             ts: Long): Future[Boolean] = {
-
-    def enqueueLogMessage() = {
-      val kafkaMessages = for {
-        vertice <- srcVertices
-        id = vertice.innerId.toIdString()
-        label <- labels
-      } yield {
-        val tsv = Seq(ts, "deleteAll", "e", id, id, label.label, "{}", GraphUtil.fromOp(dir.toByte)).mkString("\t")
-        val topic = ExceptionHandler.failTopic
-        val kafkaMsg = KafkaMessage(new ProducerRecord[Key, Val](topic, null, tsv))
-        kafkaMsg
-      }
-
-      ExceptionHandler.enqueues(kafkaMessages)
-    }
-
-    val requestTs = ts
-    val queryParams = for {
-      label <- labels
-    } yield {
-      val labelWithDir = LabelWithDirection(label.id.get, dir)
-      QueryParam(labelWithDir).limit(0, DeleteAllFetchSize).duplicatePolicy(Option(Query.DuplicatePolicy.Raw))
-    }
-
-    val step = Step(queryParams.toList)
-    val q = Query(srcVertices, Vector(step))
-
-    //    Extensions.retryOnSuccessWithBackoff(MaxRetryNum, Random.nextInt(MaxBackOff) + 1) {
-    val retryFuture = Extensions.retryOnSuccess(MaxRetryNum) {
-      fetchAndDeleteAll(q, requestTs)
-    } { case (allDeleted, deleteSuccess) =>
-      allDeleted && deleteSuccess
-    }.map { case (allDeleted, deleteSuccess) => allDeleted && deleteSuccess }
-
-    retryFuture onFailure {
-      case ex =>
-        logger.error(s"[Error]: deleteAllAdjacentEdges failed.")
-        enqueueLogMessage()
-    }
-
-    retryFuture
-  }
-
-  def flush(): Unit = clients.foreach { client =>
+  override def flush(): Unit = clients.foreach { client =>
     val timeout = Duration((clientFlushInterval + 10) * 20, duration.MILLISECONDS)
     Await.result(client.flush().toFuture, timeout)
   }
 
 
-  def createTable(zkAddr: String,
-                  tableName: String,
-                  cfs: List[String],
-                  regionMultiplier: Int,
-                  ttl: Option[Int],
-                  compressionAlgorithm: String): Unit = {
+  override def createTable(zkAddr: String,
+                           tableName: String,
+                           cfs: List[String],
+                           regionMultiplier: Int,
+                           ttl: Option[Int],
+                           compressionAlgorithm: String): Unit = {
     logger.info(s"create table: $tableName on $zkAddr, $cfs, $regionMultiplier, $compressionAlgorithm")
     val admin = getAdmin(zkAddr)
     val regionCount = admin.getClusterStatus.getServersSize * regionMultiplier
@@ -837,13 +406,64 @@ class AsynchbaseStorage(override val config: Config, vertexCache: Cache[Integer,
   }
 
 
+
+  /**
+   * Private Methods which is specific to Asynchbase implementation.
+   */
+
+
+
+
+  private def fetchKeyValuesInner(rpc: AnyRef): Deferred[util.ArrayList[KeyValue]] = {
+    rpc match {
+      case getRequest: GetRequest => client.get(getRequest)
+      case scanner: Scanner =>
+        scanner.nextRows().withCallback { kvsLs =>
+          val ls = new util.ArrayList[KeyValue]
+          if (kvsLs == null) {
+
+          } else {
+            kvsLs.foreach { kvs =>
+              if (kvs != null) kvs.foreach { kv => ls.add(kv) }
+              else {
+
+              }
+            }
+          }
+          scanner.close()
+          ls
+        }.recoverWith { ex =>
+          logger.error(s"fetchKeyValues failed.", ex)
+          scanner.close()
+          emptyKeyValues
+        }
+      case _ => Deferred.fromError(new RuntimeException(s"fetchKeyValues failed. $rpc"))
+    }
+  }
+
+  override def writeLock(rpc: HBaseRpc, expectedOpt: Option[HBaseRpc]): Future[Boolean] = {
+    val put = rpc.asInstanceOf[PutRequest]
+    val expected = expectedOpt.map(_.asInstanceOf[PutRequest].value()).getOrElse(Array.empty)
+    client(withWait = true).compareAndSet(put, expected).withCallback(ret => ret.booleanValue()).toFuture
+  }
+
+
+  private def toCacheKeyBytes(hbaseRpc: AnyRef): Array[Byte] = {
+    hbaseRpc match {
+      case getRequest: GetRequest => getRequest.key()
+      case scanner: Scanner => scanner.getCurrentKey()
+      case _ =>
+        logger.error(s"toCacheKeyBytes failed. not supported class type. $hbaseRpc")
+        Array.empty[Byte]
+    }
+  }
+
   private def getAdmin(zkAddr: String) = {
     val conf = HBaseConfiguration.create()
     conf.set("hbase.zookeeper.quorum", zkAddr)
     val conn = ConnectionFactory.createConnection(conf)
     conn.getAdmin
   }
-
   private def enableTable(zkAddr: String, tableName: String) = {
     getAdmin(zkAddr).enableTable(TableName.valueOf(tableName))
   }
@@ -864,5 +484,6 @@ class AsynchbaseStorage(override val config: Config, vertexCache: Cache[Integer,
   private def getEndKey(regionCount: Int): Array[Byte] = {
     Bytes.toBytes((Int.MaxValue / regionCount * (regionCount - 1)))
   }
+
 
 }
