+++ s2core/src/main/scala/com/kakao/s2graph/core/storage/hbase/IndexEdgeDeserializable.scala
@@ -4,7 +4,6 @@ import com.kakao.s2graph.core._
 import com.kakao.s2graph.core.mysqls.LabelMeta
 import com.kakao.s2graph.core.storage.{CanSKeyValue, StorageDeserializable, SKeyValue}
 import com.kakao.s2graph.core.types._
-import com.kakao.s2graph.core.utils.logger
 import org.apache.hadoop.hbase.util.Bytes
 
 class IndexEdgeDeserializable extends HDeserializable[IndexEdge] {
@@ -57,7 +56,20 @@ class IndexEdgeDeserializable extends HDeserializable[IndexEdge] {
 
 
   /** version 1 and version 2 is same logic */
-  override def fromKeyValues[T: CanSKeyValue](queryParam: QueryParam, _kvs: Seq[T], version: String, cacheElementOpt: Option[IndexEdge] = None): IndexEdge = {
+  override def fromKeyValues[T: CanSKeyValue](queryParam: QueryParam,
+                                              _kvs: Seq[T],
+                                              version: String,
+                                              cacheElementOpt: Option[IndexEdge] = None): IndexEdge = {
+    version match {
+      case HBaseType.VERSION4 => fromKeyValuesInnerRow(queryParam, _kvs, version, cacheElementOpt)
+      case _ => fromKeyValuesInner(queryParam, _kvs, version, cacheElementOpt)
+    }
+  }
+
+  def fromKeyValuesInner[T: CanSKeyValue](queryParam: QueryParam,
+                                          _kvs: Seq[T],
+                                          version: String,
+                                          cacheElementOpt: Option[IndexEdge] = None): IndexEdge = {
     assert(_kvs.size == 1)
 
     val kvs = _kvs.map { kv => implicitly[CanSKeyValue[T]].toSKeyValue(kv) }
@@ -106,11 +118,78 @@ class IndexEdgeDeserializable extends HDeserializable[IndexEdge] {
       if (_mergedProps.contains(LabelMeta.timeStampSeq)) _mergedProps
       else _mergedProps + (LabelMeta.timeStampSeq -> InnerVal.withLong(kv.timestamp, version))
 
-//    logger.error(s"$mergedProps")
-//    val ts = mergedProps(LabelMeta.timeStampSeq).toString().toLong
+    //    logger.error(s"$mergedProps")
+    //    val ts = mergedProps(LabelMeta.timeStampSeq).toString().toLong
 
     val ts = kv.timestamp
     IndexEdge(Vertex(srcVertexId, ts), Vertex(tgtVertexId, ts), labelWithDir, op, ts, labelIdxSeq, mergedProps)
+
+  }
+  def fromKeyValuesInnerRow[T: CanSKeyValue](queryParam: QueryParam,
+                                             _kvs: Seq[T], version: String, cacheElementOpt: Option[IndexEdge] = None): IndexEdge = {
+    assert(_kvs.size == 1)
+
+    val kvs = _kvs.map { kv => implicitly[CanSKeyValue[T]].toSKeyValue(kv) }
+
+    val kv = kvs.head
+
+    var pos = 0
+    val (srcVertexId, srcIdLen) = SourceVertexId.fromBytes(kv.row, pos, kv.row.length, version)
+    pos += srcIdLen
+    val labelWithDir = LabelWithDirection(Bytes.toInt(kv.row, pos, 4))
+    pos += 4
+    val (labelIdxSeq, isInverted) = bytesToLabelIndexSeqWithIsInverted(kv.row, pos)
+    pos += 1
+    val op = kv.row(pos)
+    pos += 1
+
+    if (pos == kv.row.length) {
+      // degree
+      val degreeVal = Bytes.toLong(kv.value)
+      val ts = kv.timestamp
+      val props = Map(LabelMeta.timeStampSeq -> InnerVal.withLong(ts, version),
+        LabelMeta.degreeSeq -> InnerVal.withLong(degreeVal, version))
+      val tgtVertexId = VertexId(HBaseType.DEFAULT_COL_ID, InnerVal.withStr("0", version))
+      IndexEdge(Vertex(srcVertexId, ts), Vertex(tgtVertexId, ts), labelWithDir, op, ts, labelIdxSeq, props)
+    } else {
+      // not degree edge
+      val index = queryParam.label.indicesMap.getOrElse(labelIdxSeq, throw new RuntimeException("invalid index seq"))
+
+      val (idxPropsRaw, endAt) = bytesToProps(kv.row, pos, version)
+      pos = endAt
+      val (tgtVertexIdRaw, tgtVertexIdLen) = if (endAt == kv.row.length) {
+        (HBaseType.defaultTgtVertexId, 0)
+      } else {
+        TargetVertexId.fromBytes(kv.row, endAt, kv.row.length, version)
+      }
+
+      val idxProps = for {
+        (seq, (k, v)) <- index.metaSeqs.zip(idxPropsRaw)
+      } yield if (k == LabelMeta.degreeSeq) k -> v else seq -> v
+
+      val idxPropsMap = idxProps.toMap
+
+      val tgtVertexId =
+        idxPropsMap.get(LabelMeta.toSeq) match {
+          case None => tgtVertexIdRaw
+          case Some(vId) => TargetVertexId(HBaseType.DEFAULT_COL_ID, vId)
+        }
+
+      val (props, _) = if (op == GraphUtil.operations("incrementCount")) {
+        val countVal = Bytes.toLong(kv.value)
+        val dummyProps = Array(LabelMeta.countSeq -> InnerVal.withLong(countVal, version))
+        (dummyProps, 8)
+      } else {
+        bytesToKeyValues(kv.value, 0, kv.value.length, version)
+      }
+
+      val _mergedProps = (idxProps ++ props).toMap
+      val mergedProps =
+        if (_mergedProps.contains(LabelMeta.timeStampSeq)) _mergedProps
+        else _mergedProps + (LabelMeta.timeStampSeq -> InnerVal.withLong(kv.timestamp, version))
+
+      val ts = kv.timestamp
+      IndexEdge(Vertex(srcVertexId, ts), Vertex(tgtVertexId, ts), labelWithDir, op, ts, labelIdxSeq, mergedProps)
+    }
   }
 }
-
