+++ s2core/src/test/scala/com/kakao/s2graph/core/Integrate/IntegrateCommon.scala
@@ -186,7 +186,7 @@ trait IntegrateCommon extends FunSuite with Matchers with BeforeAndAfterAll {
     }
     ],
     "consistencyLevel": "strong",
-    "schemaVersion": "v2",
+    "schemaVersion": "v3",
     "compressionAlgorithm": "gz",
     "hTableName": "$testHTableName"
   }"""
